+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/InstrumentedAppState.kt
@@ -39,6 +39,10 @@ public class InstrumentedAppState {
 
     private var framerateMetricsSource: FramerateMetricsSource? = null
 
+    @get:JvmName("getConfig\$internal")
+    internal var config: ImmutableConfig? = null
+        private set
+
     internal fun attach(application: Application) {
         if (this::app.isInitialized) {
             return
@@ -63,18 +67,20 @@ public class InstrumentedAppState {
     }
 
     internal fun configure(configuration: ImmutableConfig): Tracer {
+        config = configuration
         attach(configuration.application)
 
         val bootstrapSpanProcessor = spanProcessor
         val tracer = Tracer(configuration.spanEndCallbacks)
 
         configureLifecycleCallbacks(configuration)
-        app.registerComponentCallbacks(PerformanceComponentCallbacks(tracer))
 
         spanProcessor = tracer
-        spanFactory.spanProcessor = tracer
-        spanFactory.attributeLimits = configuration
-        spanFactory.networkRequestCallback = configuration.networkRequestCallback
+        spanFactory.configure(
+            tracer,
+            configuration,
+            configuration.networkRequestCallback,
+        )
 
         tracePropagationUrls = configuration.tracePropagationUrls
 
@@ -99,6 +105,12 @@ public class InstrumentedAppState {
             spanFactory.framerateMetricsSource = null
             app.unregisterActivityLifecycleCallbacks(framerateMetricsSource)
             framerateMetricsSource = null
+        }
+
+        ForegroundState.addForegroundChangedCallback { inForeground ->
+            if (!inForeground) {
+                tracer.forceCurrentBatch()
+            }
         }
 
         return tracer