+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.10.0 (2024-11-14)
+
+### Bug fixes
+
+* Untagged socket violation when StrictMode is enabled https://github.com/bugsnag/bugsnag-android-performance/issues/297
+  [#306](https://github.com/bugsnag/bugsnag-android-performance/pull/306)
+* Fixed issue where background apps didn't consistently flush the current batch of spans
+  [#305](https://github.com/bugsnag/bugsnag-android-performance/pull/305)
+
+### Enhancements
+
+* Formally introduced the `bugsnag-android-performance-compose` module and MeasuredComposable to report the performance of Composables
+  [#299](https://github.com/bugsnag/bugsnag-android-performance/pull/296)
+
 ## 1.9.1 (2024-10-30)
 
 ### Bug fixes